Adding devIocInfo.db and devLaunchEdm.sh to iocAdmin.
Add one instance of devIocInfo.db for each device to provide PVs
which allow fetching IOC_PV, IOCNAME (ioc-tst-xxx), LAUNCH_EDM,
POWER_PV, and other PVs to make it easier for scripts to find this
info given just a device prefix.
Also adds iocScreens/devLaunchEdm.sh script which fetches the path
to the edm launcher for a device and launches the edm screen.